refactor: implement home content screen with modular widgets for better structure
